Code refactoring
- Converted pure JS logic for local storage to CoffeeScript.
- Created 'autosave_token_generation.rb' file under app/models/concerns
- And handled autosave token generation logic in that file.
- Included that code in Admin and Assessor model.
